server/requirements.txt
fastapi
fastapi[standard]
sqlmodel
uvicorn
psycopg2-binary
python-dotenv
supabase
alembic
faker